[SINGLE-SHOW] update

+++ pages/api/show/[showId].js
@@ -6,7 +6,7 @@ import Shows from "../../../models/show.model";
 // Initialize the cors middleware
 const cors = initMiddleware(
   Cors({
-    methods: ["DELETE"],
+    methods: ["GET", "DELETE"],
   })
 );
 
@@ -17,6 +17,12 @@ const handler = async (req, res) => {
 
   // HANDLE DELETE
   const { showId } = req.query;
+
+  if (req.method === "GET") {
+    return Shows.findById(showId).then((show) => {
+      res.status(200).json(show);
+    });
+  }
 
   if (req.method === "DELETE") {
     return Shows.findByIdAndRemove(showId)
